Refactor GitHub commit count handling and move related functions to a new module
